requirements.txt
###############################################################################
#
# IMPORTANT:
#
# If runtime or build time dependencies are changed in this file, the committer
# *must* ensure the DEB and RPM package maintainers are informed as soon as
# possible.
#
###############################################################################

cheroot==8.*
Flask==1.*
Flask-Gravatar==0.*
Flask-Login==0.*
Flask-Mail==0.*
Flask-Migrate==2.*
Flask-SQLAlchemy==2.*
Flask-WTF==0.*
Flask-Compress==1.*
passlib==1.*
pytz==2021.*
simplejson==3.*
six==1.*
speaklater3==1.*
sqlparse==0.*
WTForms==2.*
Flask-Paranoid==0.*
psutil==5.*
psycopg2==2.8.*
python-dateutil==2.*
SQLAlchemy==1.3.*
Flask-Security-Too==3.*
bcrypt==3.*
cryptography==3.*
sshtunnel==0.*
ldap3==2.*
Flask-BabelEx==0.*
gssapi==1.6.*
Authlib==0.15.*

